+++ src/command_modules/azure-cli-interactive/azclishell/app.py
@@ -28,6 +28,7 @@ from prompt_toolkit.shortcuts import create_eventloop
 import azclishell.configuration
 from azclishell.az_lexer import AzLexer, ExampleLexer, ToolbarLexer
 from azclishell.command_tree import in_tree
+from azclishell.frequency_heuristic import DISPLAY_TIME
 from azclishell.gather_commands import add_random_new_lines
 from azclishell.key_bindings import registry, get_section, sub_section
 from azclishell.layout import create_layout, create_tutorial_layout, set_scope
@@ -119,7 +120,8 @@ class Shell(object):
 
     def __init__(self, completer=None, styles=None,
                  lexer=None, history=InMemoryHistory(),
-                 app=None, input_custom=sys.stdout, output_custom=None):
+                 app=None, input_custom=sys.stdout, output_custom=None,
+                 user_feedback=False):
         self.styles = styles
         if styles:
             self.lexer = lexer or AzLexer
@@ -137,6 +139,7 @@ class Shell(object):
         self._env = os.environ
         self.last = None
         self.last_exit = 0
+        self.user_feedback = user_feedback
         self.input = input_custom
         self.output = output_custom
         self.config_default = ""
@@ -192,7 +195,13 @@ class Shell(object):
         for _ in range(cols):
             empty_space += " "
 
-        toolbar = self._toolbar_info()
+        delta = datetime.datetime.utcnow() - START_TIME
+        if self.user_feedback and delta.seconds < DISPLAY_TIME:
+            toolbar = [
+                ' Try out the \'feedback\' command',
+                'If refreshed disappear in: {}'.format(str(DISPLAY_TIME - delta.seconds))]
+        else:
+            toolbar = self._toolbar_info()
 
         toolbar, empty_space = space_toolbar(toolbar, cols, empty_space)
         cli.buffers['bottom_toolbar'].reset(
@@ -206,7 +215,8 @@ class Shell(object):
             pass
 
         curr_cloud = "Cloud: {}".format(get_active_cloud_name())
-        tool_val = '{}'.format('Subscription: {}'.format(sub_name) if sub_name else curr_cloud)
+
+        tool_val = 'Subscription: {}'.format(sub_name) if sub_name else curr_cloud
 
         settings_items = [
             " [F1]Layout",
@@ -438,14 +448,14 @@ class Shell(object):
                 if cmd.strip() and cmd.split()[0] == 'cd':
                     handle_cd(parse_quotes(cmd))
                     continue_flag = True
-                telemetry.track_ssg('outside', cmd)
+                telemetry.track_ssg('outside', '')
 
             elif text[0] == SELECT_SYMBOL['exit_code']:
                 meaning = "Success" if self.last_exit == 0 else "Failure"
 
                 print(meaning + ": " + str(self.last_exit))
                 continue_flag = True
-                telemetry.track_ssg('exit code', cmd)
+                telemetry.track_ssg('exit code', '')
 
             elif text[0] == SELECT_SYMBOL['query']:  # query previous output
                 continue_flag = self.handle_jmespath_query(text, continue_flag)
@@ -541,6 +551,10 @@ class Shell(object):
             args = parse_quotes(cmd)
             azlogging.configure_logging(args)
 
+            if len(args) > 0 and args[0] == 'feedback':
+                SHELL_CONFIGURATION.set_feedback('yes')
+                self.user_feedback = False
+
             azure_folder = get_config_dir()
             if not os.path.exists(azure_folder):
                 os.makedirs(azure_folder)
@@ -567,6 +581,7 @@ class Shell(object):
 
             else:
                 result = self.app.execute(args)
+
             self.last_exit = 0
             if result and result.result is not None:
                 from azure.cli.core._output import OutputProducer
@@ -590,6 +605,7 @@ class Shell(object):
         from azclishell.progress import ShellProgressView
         APPLICATION.progress_controller.init_progress(ShellProgressView())
         return APPLICATION.progress_controller
+
 
     def run(self):
         """ starts the REPL """