+++ src/command_modules/azure-cli-storage/azure/cli/command_modules/storage/custom.py
@@ -15,6 +15,7 @@ from azure.cli.command_modules.storage._factory import \
     (storage_client_factory, generic_data_service_factory)
 from azure.cli.core.application import APPLICATION
 
+
 Logging, Metrics, CorsRule, \
     AccessPolicy, RetentionPolicy = get_sdk(ResourceType.DATA_STORAGE,
                                             'Logging',