+++ src/command_modules/azure-cli-interactive/HISTORY.rst
@@ -8,6 +8,12 @@ Release History
 * --progress flag
 
 
+0.3.2 (2017-05-18)
+++++++++++++++++++
+
+* Bug fixes.
+* Remove 'interactive' from completions (#3324)
+
 0.3.1 (2017-05-09)
 ++++++++++++++++++